Move getNativeBalance to swap utils
