+++ utils/e2e-tests/ts/tests/swap/evmToNative.ts
@@ -6,27 +6,7 @@ import { beforeEachWithCleanup } from "../../lib/lifecycle";
 import evmSwap from "../../lib/abis/evmSwap";
 import { decodeEventLog } from "viem";
 import * as substrate from "../../lib/substrate";
-
-type SystemAccount = {
-  data: {
-    free: bigint;
-  };
-};
-
-/// A helper function to get balance of native account.
-const getNativeBalance = async (
-  substrateApi: substrate.Api,
-  nativeAccount: string,
-) => {
-  const systemAccount = (await substrateApi.query["system"]?.["account"]?.(
-    nativeAccount,
-  )) as unknown as SystemAccount;
-
-  const free = systemAccount.data.free;
-
-  // We should explicitly convert to native bigint for future math operations.
-  return BigInt(free);
-};
+import { getNativeBalance } from "../swap/utils";
 
 const evmSwapPrecompileAddress = "0x0000000000000000000000000000000000000901";
 const bridgePotEvmAddress = "0x6d6f646c686d63732f656e310000000000000000";
