+++ utils/e2e-tests/ts/tests/swap/utils.ts
@@ -1,0 +1,24 @@
+//! Common swap utils.
+
+import * as substrate from "../../lib/substrate";
+
+type SystemAccount = {
+  data: {
+    free: bigint;
+  };
+};
+
+/// A helper function to get balance of native account.
+export const getNativeBalance = async (
+  substrateApi: substrate.Api,
+  nativeAccount: string,
+) => {
+  const systemAccount = (await substrateApi.query["system"]?.["account"]?.(
+    nativeAccount,
+  )) as unknown as SystemAccount;
+
+  const free = systemAccount.data.free;
+
+  // We should explicitly convert to native bigint for future math operations.
+  return BigInt(free);
+};
